forwarding forwarded messages

+++ utils.py
@@ -35,9 +35,18 @@ class MyTranslator(app_commands.Translator):
                 return 'Удалить сохраненное сообщение(я)'
         return
     
-async def create_send_embeds(ctx: discord.Interaction, messages: list[discord.Message], show_original: bool=True, anonymous: bool=False, show_ids: bool=False) -> dict:
+async def create_send_embeds(ctx: discord.Interaction, messages: list[discord.Message | discord.Embed], show_original: bool=True, anonymous: bool=False, show_ids: bool=False) -> dict:
     embeds: list[discord.Embed] = []
     for i, message in enumerate(messages):
+        if type(message) is discord.Embed:
+            if i == 0:
+                message.title = f'{getenv("EMOJI") or ""} *Forwarded*' if not ctx.locale is discord.Locale.russian else f'{getenv("EMOJI") or ""} *Переслано*'
+            else:
+                message.title = None
+            if not show_original and not message.fields[-1].inline:
+                message.remove_field(-1)
+            embeds.append(message)
+            continue
         tenor = message.embeds and message.embeds[0].url is not None and message.embeds[0].url.startswith('https://tenor.com/view/') # kill tenor
         image = discord.utils.find(lambda a: a.content_type in image_types, message.attachments)
         if i == 0:
